requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pydantic>=2.0.0
requests>=2.31.0
Pillow>=10.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
